requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
gunicorn==23.0.0
Flask==3.1.0
openai==1.69.0
pandas==2.2.3
python-dotenv==1.1.0
matplotlib==3.10.1
pdf2image==1.16.3
Pillow==10.0.1
requests==2.31.0
PyPDF2==3.0.1